Correct PinholeCalibrate to work with new RANSAC

+++ src/vw/Camera/PinholeModelCalibrate.h
@@ -407,33 +407,33 @@ public:
 /// \parm lm_iter number of Levenberg Marquardt iterations to run at every RANSAC iteration
 /// \return indices of points and pixels RANSAC considers to be inliers
 template<class Ex1T, class Ex2T, class Ex3T, class Ex4T>
-std::vector<int> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
+std::vector<size_t> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
   PinholeModelRansacFitting<Ex1T, Ex2T, Ex3T, Ex4T> fitting(m, lm_iter);
   PinholeModelRansacError error;
 
   math::RandomSampleConsensus<PinholeModelRansacFitting<Ex1T, Ex2T, Ex3T, Ex4T>, PinholeModelRansacError> ransac(fitting, error, inlier_threshold);
   m = ransac(pixels, points, ransac_iter);
 
-  std::vector<int> inlier_indices(ransac.inlier_indices(m, pixels, points));
+  std::vector<size_t> inlier_indices(ransac.inlier_indices(m, pixels, points));
   vw_out(DebugMessage, "Camera") << "RANSAC classified as inliers: " << inlier_indices.size() << '/' << pixels.size() << std::endl;
   return inlier_indices;
 }
 
 /// Convenience overload with less template parameters
 template<class Ex1T, class Ex2T, class Ex3T>
-inline std::vector<int> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
+inline std::vector<size_t> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
   return pinholemodel_calibrate_ransac<Ex1T, Ex2T, Ex3T, PinholeModelSerializeNull>(m, pixels, points, inlier_threshold, ransac_iter, lm_iter);
 }
 
 /// Convenience overload with less template parameters
 template<class Ex1T, class Ex2T>
-inline std::vector<int> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
+inline std::vector<size_t> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
   return pinholemodel_calibrate_ransac<Ex1T, Ex2T, PinholeModelSerializeNull>(m, pixels, points, inlier_threshold, ransac_iter, lm_iter);
 }
 
 /// Convenience overload with less template parameters
 template<class Ex1T>
-inline std::vector<int> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
+inline std::vector<size_t> pinholemodel_calibrate_ransac(PinholeModel& m, const std::vector<Vector2>& pixels, const std::vector<Vector3>& points, double inlier_threshold = 10, int ransac_iter = 0, int lm_iter = 100) {
   return pinholemodel_calibrate_ransac<Ex1T, PinholeModelSerializeNull>(m, pixels, points, inlier_threshold, ransac_iter, lm_iter);
 }
